+++ app/javascript/entrypoints/v_posts.ts
@@ -3,9 +3,10 @@
 import E621Type from "@/interfaces/E621";
 declare const E621: E621Type;
 
-import "@/pages/posts/SearchFilters";
 import PostModeMenu from "@/pages/posts/post_mode_menu";
 import "@/pages/posts/post_search";
 import Post from "@/pages/posts/posts";
+import "@/pages/posts/SearchControls";
+import "@/pages/posts/SearchFilters";
 
 E621.Registry.register("v_posts", { Post, PostModeMenu });
