+++ app/javascript/src/js/utility/SVGIcon.ts
@@ -33,8 +33,6 @@ export default class SVGIcon {
     "times": '<path d="M18 6 6 18"/><path d="m6 6 12 12"/>',
     "hash": '<line x1="4" x2="20" y1="9" y2="9"/><line x1="4" x2="20" y1="15" y2="15"/><line x1="10" x2="8" y1="3" y2="21"/><line x1="16" x2="14" y1="3" y2="21"/>',
     "shuffle": '<path d="M2 18h1.4c1.3 0 2.5-.6 3.3-1.7l6.1-8.6c.7-1.1 2-1.7 3.3-1.7H22"/><path d="m18 2 4 4-4 4"/><path d="M2 6h1.9c1.3 0 2.5.6 3.3 1.7l2.1 2.9"/><path d="m18 14 4 4-4 4"/><path d="M21.7 17.4h-.3c-1.3 0-2.5-.6-3.3-1.7L15 14.5"/>',
-    "sidebar_hide": '<rect x="2" y="3" width="20" height="18" rx="2"/><path d="M11 3v18"/><path d="m8 9-3 3 3 3"/>',
-    "sidebar_show": '<rect x="2" y="3" width="20" height="18" rx="2"/><path d="M11 3v18"/><path d="m5 9 3 3-3 3"/>',
     "pencil": '<path d="M21.174 6.812a1 1 0 0 0-3.986-3.987L3.842 16.174a2 2 0 0 0-.5.83l-1.321 4.352a.5.5 0 0 0 .623.622l4.353-1.32a2 2 0 0 0 .83-.497z"/><path d="m15 5 4 4"/>',
 
     "clock_fading": '<path d="M12 2a10 10 0 0 1 7.38 16.75"/><path d="M12 6v6l4 2"/><path d="M2.5 8.875a10 10 0 0 0-.5 3"/><path d="M2.83 16a10 10 0 0 0 2.43 3.4"/><path d="M4.636 5.235a10 10 0 0 1 .891-.857"/><path d="M8.644 21.42a10 10 0 0 0 7.631-.38"/>',
